스킬시스템 구현중중

+++ Assets/Script/SkillScript/ISkill.cs
@@ -4,6 +4,6 @@ public interface ISkill
 {
     public int SkillType { get; set; }
     public int SkillDamage { get; set;}
-    public int SkillCoolTime { get; set; }
+    public float SkillCoolTime { get; set; }
     void UseSkill();
 }
